Extract text file loading functions to a separate .cpp file, so that they can be used in test automation as well
